possibly bad merged

+++ db/migrate/20160404162153_remove_student_school_years_without_students.rb
@@ -1,8 +1,10 @@
 class RemoveStudentSchoolYearsWithoutStudents < ActiveRecord::Migration[4.2]
   def change
-    StudentSchoolYear.find_each do |student_school_year|
-      student_school_year.destroy if student_school_year.student.nil?
-      student_school_year.destroy if student_school_year.school_year.nil?
+    if defined? StudentSchoolYear
+      StudentSchoolYear.find_each do |student_school_year|
+        student_school_year.destroy if student_school_year.student.nil?
+        student_school_year.destroy if student_school_year.school_year.nil?
+      end
     end
   end
 end
